moved page components under pages
