--- src/components/Login.tsx
@@ -1,5 +1,0 @@
-import React from 'react'
-
-export default function Login() {
-    return <h2>Login</h2>;
-}
