--- src/components/Signup.tsx
@@ -1,5 +1,0 @@
-import React from "react"
-
-export default function Signup() {
-    return <h2>Signup</h2>;
-}
