--- src/components/UserBoard.tsx
@@ -1,5 +1,0 @@
-import React from "react"
-
-export default function UserBoard() {
-    return <h2>UserBoard</h2>;
-}
