+++ src/pages/Login.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+/**
+ * @return {JSX.Element}
+ */
+export default function Login() {
+  return <h2>Login</h2>;
+}
