+++ src/pages/Signup.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+/**
+ * @return {JSX.Element}
+ */
+export default function Signup() {
+  return <h2>Signup</h2>;
+}
