+++ src/pages/UserBoard.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+/**
+ * @return {JSX.Element}
+ */
+export default function UserBoard() {
+  return <h2>UserBoard</h2>;
+}
